feat: worker适配
