--- src/main/workers/index.ts
@@ -1,21 +1,0 @@
-import PromiseWorker from 'promise-worker';
-import CodingWorker from './coding.worker.ts';
-
-export class BasicWorker extends PromiseWorker {
-  private worker: Worker;
-
-  constructor(worker: Worker) {
-    super(worker);
-    this.worker = worker;
-  }
-
-  terminate() {
-    this.worker.terminate();
-  }
-}
-
-export class CodingPromiseWorker extends BasicWorker {
-  constructor() {
-    super(new CodingWorker());
-  }
-}
