+++ src/renderer/services/quotation.ts
@@ -1,7 +1,6 @@
 import NP from 'number-precision';
+import { fromUint8Array } from 'js-base64';
 import request from '@/utils/request';
-
-const { base64 } = window.contextModules;
 
 export async function GetQuotationsFromEastmoney() {
   try {
@@ -1335,7 +1334,7 @@ export async function GetShanghaiGoldGoodsFromEastmoney() {
 export async function GetMainFundFromEastmoney(code: string) {
   try {
     const { rawBody } = await request(`https://webquotepic.eastmoney.com/GetPic.aspx?nid=${code}&imageType=FFRS1&type=FFR`);
-    const b64encoded = base64.fromUint8Array(new Uint8Array(rawBody));
+    const b64encoded = fromUint8Array(new Uint8Array(rawBody));
     return `data:image/png;base64,${b64encoded}`;
   } catch (error) {
     return;
